back/pyproject.toml
[project]
name = "minilysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.8.1",
    "fastapi[standard]>=0.115.8",
    "tortoise-orm[asyncpg]>=0.24.0",
]

[tool.aerich]
tortoise_orm = "db.session.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
